Move the lock file utils to shared
The shared package contains the lockfile utils, so that they can be shared accross apps
